backend/pyproject.toml
[project]
name = "smartrade-adk"
version = "0.1.0"
description = "Google ADK based smartrade agent"
requires-python = ">=3.12"
dependencies = [
    "ag-ui-adk>=0.3.2",
    "bs4>=0.0.2",
    "google-adk>=1.18.0",
    "httpx>=0.28.1",
    "langchain-community>=0.4.1",
    "litellm>=1.79.1",
    "matplotlib>=3.10.7",
    "mplfinance>=0.12.10b0",
    "pandas>=2.3.3",
]
